initial commit
This project seems to work ok.  It is a 2D platformer with a moving
platform that follows an arbitrary GameMaker path.  There is a bit of a
shimmying problem when the player lands on the platform, most likely due
to a rounding error in the displacement of the player due to stepwise
motion of the platform.